// src/app/shared/models/client.ts
export class Client {
  id?: number;
  firstName: string;
  lastName: string;
  middleName?: string;
  sex: string;
  dateOfBirth: string;
  marital: number;
  education: number;
  occupation: string;
  address1: string;
  address2?: string;
  city: string;
  state: string;
  zip: string;
  ssn?: string;
  license: string;
}